[FIXUP] "crew" AI would never re-task to newly activated objectives.
AI are unable to move anywhere once they have mounted a static weapon at site X. If the AI subsytem determines site Y is now higher priority (needs AI), and site Y is within the reallocation range (600m), the subsystem will deallocate the AI from site X and give them a move order to site Y.

But "crew" AI are in static weapons, and can't move anywhere!

So when a "crew" objective is deactivated we need to `moveOut` any AI on the statics so they are free to move somewhere else.